Refactoring
- Moved AstInterpreter from AST to Evaluating
- Combined Parser constructor and Parse method in static method
